# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for Star Mall service.",
        "title": "Star Mall API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9090",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/inventory/create": {
            "put": {
                "description": "创建库存",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "创建库存",
                "parameters": [
                    {
                        "description": "inventory",
                        "name": "inventory",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.InventoryCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "inventory created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "invalid token claims",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/inventory/search/{MerchantID}": {
            "get": {
                "description": "获取库存",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "GetInventoryByMerchant 获取库存",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "inventory id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "inventory 数组",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Inventory"
                            }
                        }
                    },
                    "401": {
                        "description": "invalid token claims",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "inventory not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/inventory/update": {
            "patch": {
                "description": "更新库存",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "更新库存",
                "parameters": [
                    {
                        "description": "inventory",
                        "name": "inventory",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.InventoryCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "inventory created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "invalid token claims",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/inventory/{id}": {
            "get": {
                "description": "获取库存",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "GetInventory 获取库存",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "inventory id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "inventory",
                        "schema": {
                            "$ref": "#/definitions/domain.Inventory"
                        }
                    },
                    "401": {
                        "description": "invalid token claims",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "inventory not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Inventory": {
            "type": "object",
            "properties": {
                "availableStock": {
                    "description": "可用库存",
                    "type": "integer"
                },
                "createAt": {
                    "type": "integer"
                },
                "lowStockThreshold": {
                    "description": "低库存阈值",
                    "type": "integer"
                },
                "productID": {
                    "type": "integer"
                },
                "reservedStock": {
                    "description": "已锁定库存",
                    "type": "integer"
                },
                "updateAt": {
                    "type": "integer"
                },
                "version": {
                    "description": "乐观锁版本",
                    "type": "integer"
                }
            }
        },
...
        "handler.InventoryCreateRequest": {
            "type": "object",
            "properties": {
                "count": {
                    "description": "@Describe: 库存数量",
                    "type": "integer"
                },
                "low_stock_threshold": {
                    "description": "@Describe: 低库存阈值",
                    "type": "integer"
                },
                "product_id": {
                    "description": "@Describe: 商品ID",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "输入 JWT token，格式为：Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}